requirements.txt
apache-airflow==2.10.5
apache-airflow-providers-amazon==9.4.0
boto3==1.37.20
awscli==2.25.3